# apps/api/openapi/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/v1/offers/{id}": {
      "get": {
        "operationId": "OffersController_getById",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Offer detail with business and location"
          }
        },
        "summary": "Get offer detail",
        "tags": [
          "Offers"
        ]
      },
      "patch": {
        "operationId": "OffersController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Offer updated"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Update an offer (admin or business owner)",
        "tags": [
          "Offers"
        ]
      },
      "delete": {
        "operationId": "OffersController_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Offer deactivated"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Deactivate an offer (admin or business owner)",
        "tags": [
          "Offers"
        ]
      }
    },
...
    "/api/v1/businesses/public/{id}": {
      "get": {
        "operationId": "BusinessesPublicController_storefront",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Public business with locations and hours"
          }
        },
        "summary": "Get a business storefront: the business plus its active locations and weekly hours",
        "tags": [
          "Businesses"
        ]
      }
    },
...
    "/api/v1/reviews": {
      "post": {
        "operationId": "ReviewsController_create",
        "parameters": [],
        "responses": {
          "201": {
            "description": "Review created"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Create a review for a completed order",
        "tags": [
          "Reviews"
        ]
      }
    },
    "/api/v1/businesses/public/{businessId}/reviews": {
      "get": {
        "operationId": "ReviewsFeedsController_listBusinessReviews",
        "parameters": [
          {
            "name": "businessId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Paginated reviews of a public business"
          }
        },
        "summary": "Public review feed of a business (visible reviews only)",
        "tags": [
          "Reviews"
        ]
      }
    },
    "/api/v1/offers/{id}/reviews": {
      "get": {
        "operationId": "ReviewsFeedsController_listOfferReviews",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Paginated reviews of a public offer"
          }
        },
        "summary": "Public review feed of an offer (visible only)",
        "tags": [
          "Reviews"
        ]
      }
    },
    "/api/v1/me/reviews": {
      "get": {
        "operationId": "ReviewsFeedsController_listMyReviews",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Paginated reviews written by the caller"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "My review history, including reviews hidden from the public",
        "tags": [
          "Reviews"
        ]
      }
    },
...
    "/api/v1/favorites": {
      "get": {
        "operationId": "FavoritesController_list",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Paginated favorites with the offer embedded"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "List my favorites",
        "tags": [
          "Favorites"
        ]
      },
      "post": {
        "operationId": "FavoritesController_add",
        "parameters": [],
        "responses": {
          "201": {
            "description": "Favorite created or already present"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Add a favorite (idempotent)",
        "tags": [
          "Favorites"
        ]
      }
    },
    "/api/v1/favorites/{offerId}": {
      "delete": {
        "operationId": "FavoritesController_remove",
        "parameters": [
          {
            "name": "offerId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Favorite removed"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Remove one of my favorites by offer id",
        "tags": [
          "Favorites"
        ]
      }
    },
...
  },
  "info": {
    "title": "Role API",
    "description": "Backend API para Role — marketplace de comida excedente",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
...
  }
}